Refactored code so that all purchasing and item checking goes through ItemStoreFront
This will allow it to monitor all purchases that happen, allowing users of the class to
be notified of changes in the shop. Previously, purchases circumvented the ItemStoreFront
by modifying its item stock directly. This made it difficult for us to notify the shop as
a whole that a purchase had occurred (and the UI may need to be updated).

I undid many of the changes to add observable fields from the previous commit since they
are no longer necessary.